// export/types/src/contracts/IGovernorINNImpl.ts
export declare namespace IGovernorINN {
  export type ProposalRequestStruct = {
    offchainID: BytesLike;
    proposalType: BigNumberish;
    actionType: BigNumberish;
    description: string;
    data: BytesLike;
  };

  export type ProposalRequestStructOutput = [
    string,
    number,
    number,
    string,
    string
  ] & {
    offchainID: string;
    proposalType: number;
    actionType: number;
    description: string;
    data: string;
  };
}
// ...
export interface NewValidatorProposalCreatedEventObject {
  proposalID: string;
  proposer: string;
  validatorEOA: string;
  offchainID: string;
  description: string;
  validatorName: string;
}